refactor: separate some concerns and tidy up
