+++ default-plugins/compact-bar/src/clipboard_utils.rs
@@ -1,0 +1,28 @@
+use zellij_tile::prelude::*;
+use crate::LinePart;
+
+pub fn text_copied_hint(copy_destination: CopyDestination) -> LinePart {
+    let hint = match copy_destination {
+        CopyDestination::Command => "Text piped to external command",
+        #[cfg(not(target_os = "macos"))]
+        CopyDestination::Primary => "Text copied to system primary selection",
+        #[cfg(target_os = "macos")] // primary selection does not exist on macos
+        CopyDestination::Primary => "Text copied to system clipboard",
+        CopyDestination::System => "Text copied to system clipboard",
+    };
+    LinePart {
+        part: serialize_text(&Text::new(&hint).color_range(2, ..).opaque()),
+        len: hint.len(),
+        tab_index: None,
+    }
+}
+
+pub fn system_clipboard_error() -> LinePart {
+    let hint = " Error using the system clipboard.";
+    LinePart {
+        part: serialize_text(&Text::new(&hint).color_range(2, ..).opaque()),
+        len: hint.len(),
+        tab_index: None,
+    }
+}
+
